pyproject.toml
[tool.poetry]
name = "mix-trap"
version = "0.1.0"
description = ""
authors = ["Louis Sanna <louis.sanna@ekimetrics.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
mistralai = "^0.2.0"
python-dotenv = "^1.0.1"
requests = "^2.32.2"
newsapi-python = "^0.2.7"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
